requirements.txt
appnope==0.1.2
argon2-cffi==20.1.0
asgiref==3.3.4
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
decorator==5.0.5
defusedxml==0.7.1
Django==3.2
django-cors-headers==3.7.0
django-environ==0.4.5
django-health-check==3.16.3
entrypoints==0.3
gunicorn==20.1.0
idna==2.10
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
MarkupSafe==1.1.1
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
nltk==3.5
notebook==6.3.0
numpy==1.20.1
packaging==20.9
pandas==1.2.3
pandocfilters==1.4.3
parso==0.8.2
pdfminer.six==20201018
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.10.0
prompt-toolkit==3.0.18
psutil==5.8.0
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.8.1
PyMuPDF==1.18.9
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
pyzmq==22.0.3
regex==2020.11.13
requests==2.25.1
scikit-learn==0.24.1
scipy==1.6.1
Send2Trash==1.5.0
six==1.15.0
sortedcontainers==2.3.0
soupsieve==2.2.1
sqlparse==0.4.1
terminado==0.9.4
testpath==0.4.4
threadpoolctl==2.1.0
tornado==6.1
tqdm==4.59.0
traitlets==5.0.5
urllib3==1.26.4
wcwidth==0.2.5
webencodings==0.5.1
